Conflicts:
	iOS/tests/ConnectionIssues/pubnub/libs/PubNub/Misc/PNMacro.h
	iOS/tests/ConnectionIssues1/pubnub/libs/PubNub/Misc/PNMacro.h
	iOS/tests/ConnectionResume/pubnub/libs/PubNub/Misc/PNMacro.h
	iOS/tests/IdleTimeout/pubnub/libs/PubNub/Misc/PNMacro.h
	iOS/tests/JSONAbsent/pubnub/libs/PubNub/Misc/PNMacro.h
	iOS/tests/JSONPresent/pubnub/libs/PubNub/Misc/PNMacro.h
	iOS/tests/PubNubTestBackground/pubnub/libs/PubNub/Misc/PNMacro.h
	iOS/tests/PubNubTestBackgroundTime/pubnub/libs/PubNub/Misc/PNMacro.h
	iOS/tests/Subscription/pubnubBackground/libs/PubNub/Misc/PNMacro.h
	iOS/tests/ThreadBlocked/pubnub/libs/PubNub/Misc/PNMacro.h
	iOS/tests/iOSUnitTests/pubnub.xcodeproj/xcshareddata/xcschemes/pubnubTestsDebug.xcscheme
	iOS/tests/iOSUnitTests/pubnub/libs/PubNub/Misc/PNMacro.h
	iOS/tests/iOSUnitTests/pubnubTests/Configuration.m
	iOS/tests/pubnubBackground/pubnubBackground/libs/PubNub/Misc/PNMacro.h

+++ iOS/tests/JSONAbsent/pubnub/libs/PubNub/Data/PNAccessRightOptions.m
@@ -124,7 +124,7 @@ static NSUInteger const kPNDefaulfAccessPeriodDuration = 5;
         [description appendFormat:@" users: %@;", self.clientsAuthorizationKeys];
     }
 
-    [description appendFormat:@" access period duration: %d>", self.accessPeriodDuration];
+    [description appendFormat:@" access period duration: %lu>", (unsigned long)self.accessPeriodDuration];
 
 
     return description;
